+++ DellOpenManage/Classes/Schedule.psm1
@@ -1,4 +1,6 @@
 Class Schedule {
+    [Boolean]$RunNow
+    [Boolean]$RunLater
     [nullable[DateTime]]$StartTime
     [nullable[DateTime]]$EndTime
     [String]$Cron