+++ Tests/Device.Tests.ps1
@@ -3,17 +3,45 @@ Connect-OMEServer -Name $Global:OMEServer -Credentials $credentials -IgnoreCerti
 Describe "Device Tests" {
     BeforeEach {
         $TestDeviceHosts = @("100.79.6.22")
+        $TestDeviceHostsNew = @("100.79.6.64")
         $TestDeviceServiceTags = @("37KP0Q2")
+        $TestDiscoveryJobName = "TestDiscovery01"
         $TestDeviceModel = "PowerEdge R740"
         $TestGroup = "Dell iDRAC Servers"
         $TestiDRACUsername = $Global:iDRACUsername
         $TestiDRACPassword = $Global:iDRACPassword
     }
-    Context "Checks 1" {
-        It "Should return a discovery job id" {
-            New-OMEDiscovery -Hosts $TestDeviceHosts -DiscoveryUserName $TestiDRACUsername -DiscoveryPassword $(ConvertTo-SecureString $TestiDRACPassword -AsPlainText -Force) -Wait -Verbose | Should -BeGreaterThan 0 
+    Context "Discovery Job Checks" {
+        It "NewDiscoveryJob > Should return a discovery job id" {
+            New-OMEDiscovery -Name $TestDiscoveryJobName -Hosts $TestDeviceHosts -DiscoveryUserName $TestiDRACUsername -DiscoveryPassword $(ConvertTo-SecureString $TestiDRACPassword -AsPlainText -Force) -Wait | Should -BeGreaterThan 0
         }
 
+        It "GetDiscoveryJob > Should return 1 discovery job" {
+            $TestDiscoveryJobName | Get-OMEDiscovery | Measure-Object | Select-Object -ExpandProperty Count | Should -Be 1
+        }
+
+        It "EditDiscoveryJobAppend > Should return 2 targets hosts" {
+            $TestDiscoveryJobName | Get-OMEDiscovery | Edit-OMEDiscovery -Hosts $TestDeviceHostsNew -Mode "Append" -DiscoveryUserName $TestiDRACUsername -DiscoveryPassword $(ConvertTo-SecureString $TestiDRACPassword -AsPlainText -Force)
+            $TestDiscoveryJobName | Get-OMEDiscovery | Select-Object -ExpandProperty Hosts | Measure-Object | Select-Object -ExpandProperty Count | Should -Be 2
+        }
+
+        It "EditDiscoveryJobRemove > Should return 1 target hosts" {
+            $TestDiscoveryJobName | Get-OMEDiscovery | Edit-OMEDiscovery -Hosts $TestDeviceHostsNew -Mode "Remove" -DiscoveryUserName $TestiDRACUsername -DiscoveryPassword $(ConvertTo-SecureString $TestiDRACPassword -AsPlainText -Force)
+            $TestDiscoveryJobName | Get-OMEDiscovery | Select-Object -ExpandProperty Hosts | Measure-Object | Select-Object -ExpandProperty Count | Should -Be 1
+        }
+
+        It "EditDiscoveryJobKeepHosts > Should return 1 target hosts" {
+            $TestDiscoveryJobName | Get-OMEDiscovery | Edit-OMEDiscovery -DiscoveryUserName $TestiDRACUsername -DiscoveryPassword $(ConvertTo-SecureString $TestiDRACPassword -AsPlainText -Force)
+            $TestDiscoveryJobName | Get-OMEDiscovery | Select-Object -ExpandProperty Hosts | Measure-Object | Select-Object -ExpandProperty Count | Should -Be 1
+        }
+
+        It "EditDiscoveryJobRunLater > Should return matching cron string" {
+            $TestCron = "0 0 0 ? * sun *"
+            $TestDiscoveryJobName | Get-OMEDiscovery | Edit-OMEDiscovery -Schedule "RunLater" -ScheduleCron $TestCron -DiscoveryUserName $TestiDRACUsername -DiscoveryPassword $(ConvertTo-SecureString $TestiDRACPassword -AsPlainText -Force)
+            $TestDiscoveryJobName | Get-OMEDiscovery | Select-Object -ExpandProperty Schedule | Select-Object -ExpandProperty Cron | Should -Be $TestCron
+        }
+    }
+    Context "Device Checks" {
         It "Should return ALL Devices" {
             Get-OMEDevice | Measure-Object | Select-Object -ExpandProperty Count | Should -BeGreaterThan 0
         }
@@ -30,8 +58,13 @@ Describe "Device Tests" {
             "1000" | Get-OMEDevice -FilterBy "Type" | Measure-Object | Select-Object -ExpandProperty Count | Should -BeGreaterThan 0
         }
 
+        It "Should return device inventory" {
+            $TestDeviceServiceTags | Get-OMEDevice | Get-OMEDeviceDetail -InventoryType "deviceSoftware" | Measure-Object | Select-Object -ExpandProperty Count | Should -BeGreaterThan 0
+        }
+    }
+    Context "Group Checks" {
         It "Should return ALL Groups" {
-            Get-OMEGroup | Select-Object -ExpandProperty Name | Should -BeGreaterThan 2
+            Get-OMEGroup | Measure-Object | Select-Object -ExpandProperty Count | Should -BeGreaterThan 2
         }
 
         It "Should return a specific Group" {
@@ -40,10 +73,6 @@ Describe "Device Tests" {
 
         It "Should return Devices by Group" {
             $TestGroup | Get-OMEGroup | Get-OMEDevice | Measure-Object | Select-Object -ExpandProperty Count | Should -BeGreaterThan 0
-        }
-
-        It "Should return device inventory" {
-            $TestDeviceServiceTags | Get-OMEDevice | Get-OMEDeviceDetail -InventoryType "software" | Measure-Object | Select-Object -ExpandProperty Count | Should -BeGreaterThan 0
         }
     }
 }